requirements.txt
pyyaml
flask
flask-cors
httpx
pydantic
sentence-transformers

# Optional providers (choose any)
#google-api-python-client